Change report and files
