--- Mukhin/lab3/Source/lab3.h
@@ -1,16 +1,0 @@
-#pragma once
-#include <cstdlib>
-#include <fstream>
-
-template <class type>
-class Stack{
-    private:
-        type* st;
-        size_t len;
-    public:
-        Stack(size_t max_size);
-        ~Stack();
-        void push(type elem);
-        type pop();
-        bool isEmpty();
-};
