+++ Mukhin/lab3/Source/lab3.hpp
@@ -1,5 +1,18 @@
-#include "lab3.h"
 #include <fstream>
+#include <cstdlib>
+
+template <class type>
+class Stack{
+private:
+    type* st;
+    size_t len;
+public:
+    Stack(size_t max_size);
+    ~Stack();
+    void push(type elem);
+    type pop();
+    bool isEmpty();
+};
 
 template <class type>
 Stack<type>::Stack(size_t max_size) {
